fix: topic handling in client & now clean shutdown of ConnectWebsocket
When using batch subscriptions, and later on trying to unsubscribe
from only one of the topics in the batch, the topic would not be
found in the _topics list of ConnectWebsocket. This is now handled
better.

The second issue was that ConnectWebsocket always produced a long
stack trace during shutdown because of an unhandled exeception in
the _run method/task. This is also fixed now.